refactor: add util for validate uuid
